fix: extract escapeHtml and sanitize disease names
